donut-specific code divided into 2 files based on dependencies, removed excess #include
