+++ examples/transform_examples/donut_example.cpp
@@ -1,14 +1,14 @@
 
 #include "../../modules/transform/distance_potential_minimization.hpp"
-#include "donuts.hpp"
-#include <iostream>
 
+#include "donuts_test_io.hpp"
+#include "donuts.hpp"
+
+#include <iostream>
 #include <filesystem>
 
-
-//#include "../../modules/utils/image_processing/image_filter.hpp" // for only test call of imfilter,TODO remove
-#include "assets/helpers.cpp"
-#include "modules/utils/visualizer.hpp" // for only blaze2bmp_norm
+//#include "assets/helpers.cpp" // only for csv reader called in debug f() function
+//#include "modules/utils/visualizer.hpp" // for only blaze2bmp_norm
 
 
 
@@ -20,7 +20,7 @@ void filter_donut(std::string filename) {
 
     auto mask = weightingMask<double>(donut.rows(), donut.columns(), donut.columns()/3, 6);
     donut = mask % donut;
-    vector2bmp(matrix2vv(donut), filename + "filtered.bmp");
+    vector2bmp(matrix2vv(donut), filename + ".filtered.bmp");
 }
 
 
@@ -40,7 +40,7 @@ void fit_donut(std::string filename)
     auto donut_input = donut;
 
     donut = radial_diff(donut);
-    vector2bmp(matrix2vv(donut), filename + "filtered.bmp");
+    vector2bmp(matrix2vv(donut), filename + ".filtered.bmp");
 
     //debug_out = blaze::submatrix(donut, 0, 0, 10, 10);
     //std::cout << "filtered:\n" << debug_out << "\n";
@@ -49,12 +49,12 @@ void fit_donut(std::string filename)
     auto mask = weightingMask<double>(donut.rows(), donut.columns(), donut.columns()*0.4, 15); // new donuts
     donut = mask % donut;
 
-    vector2bmp(matrix2vv(mask), filename + "mask_outer.bmp");
+    vector2bmp(matrix2vv(mask), filename + ".mask_outer.bmp");
 
     //mask = weightingMask<double>(128, 128, 10, 2);
     //vector2bmp(matrix2vv(mask), "mask_inner.bmp");
 
-    vector2bmp(matrix2vv(donut), filename + "mask_applied.bmp");
+    vector2bmp(matrix2vv(donut), filename + ".mask_applied.bmp");
 
     auto [u, v] = metric::DPM_detail::gvf(donut, 0.1, 1, 10);
     vector2bmp(matrix2vv(u), filename + ".u.bmp");
@@ -87,7 +87,7 @@ void fit_donut(std::string filename)
 
     auto t1 = std::chrono::steady_clock::now();
 
-    auto result = metric::fit_hysteresis(donut, init_x, init_y, init_r, steps, sigma);//, 0.2, 1e-8); // TODO remove non-default thresold
+    auto result = metric::fit_hysteresis(donut, init_x, init_y, init_r, steps, sigma);//, 0.2, 1e-8);
 
     auto t2 = std::chrono::steady_clock::now();
 
@@ -137,7 +137,7 @@ void fit_donut(std::string filename)
 }
 
 
-//* single file provided in the repo
+//* single file
 
 int main() {
     //fit_donut("assets/donuts/crop/crop_2020-07-27_16_23_01_776_donut1_128.png");
@@ -149,41 +149,6 @@ int main() {
 
 //*/
 
-
-
-/* single file
-
-int main() {
-    fit_donut("test01.png");
-    return 0;
-}
-
-//*/
-
-
-/* blur test
-
-int main() {
-
-//    blaze::DynamicMatrix<double> a = { {0.5, 0.5, 0.5, 0.5, 0.5},
-//                                       {0.5, 0.8, 0.5, 0.5, 0.5},
-//                                       {0.5, 0.5, 0.5, 0.5, 0.5},
-//                                       {0.5, 0.5, 0.5, 0.5, 0.5},
-//                                       {0.5, 0.5, 0.5, 0.5, 0.5}
-//                                     };
-    blaze::DynamicMatrix<double> a = { {0.5, 0.5, 0.5, 0.5},
-                                       {0.5, 0.8, 0.5, 0.5},
-                                       {0.5, 0.5, 0.5, 0.5},
-                                       {0.5, 0.5, 0.5, 0.5},
-                                       {0.5, 0.5, 0.5, 0.5}
-                                     };
-
-    std::cout << radial_diff(a) << "\n";
-
-    return 0;
-}
-
-//*/
 
 
 
@@ -276,399 +241,399 @@ int main() {
 
 
 
-// fit_ellipse by parts
+//// fit_ellipse by parts - for debugging
 
-int f() {
-//int main() {
+//int f() { // function for debug purpose
+////int main() {
 
-//    auto donut = read_png_donut<double>("assets/donuts/crop/crop_2020-07-27_16_23_01_776_donut1_128.png");
-//    donut = radial_diff(donut);
+////    auto donut = read_png_donut<double>("assets/donuts/crop/crop_2020-07-27_16_23_01_776_donut1_128.png");
+////    donut = radial_diff(donut);
 
+////    auto mask = weightingMask<double>(donut.rows(), donut.columns(), donut.columns()/3, 6);
+////    donut = mask % donut;
+
+//    /* // generate & save ring
+//    auto ring = weightingMask<double>(128 - 7, 128 - 7, 40, 6);
+//    blaze::DynamicMatrix<double> donut (128, 128, 0);
+//    blaze::submatrix(donut, 7, 7, 128 - 7, 128 - 7) = ring;
+//    blaze_dm_to_csv(donut, "ring.csv");
+//    vector2bmp(matrix2vv(donut), "ring.bmp");
+//    // */
+
+//    auto donut = read_csv_blaze<double>("test01.csv", ",");
+
+
+//    //vector2bmp(matrix2vv(mask), "mask_outer.bmp");
+
+//    //mask = weightingMask<double>(128, 128, 10, 2);
+//    //vector2bmp(matrix2vv(mask), "mask_inner.bmp");
+
+//    //vector2bmp(matrix2vv(donut), "mask_applied.bmp");
+
+
+//    //auto donut = read_png_donut<double>("assets/donuts/crop/crop_2020-07-27_16_23_01_776_donut1.png");
+
+//    //auto donut = read_png_donut<double>("assets/donuts/crop/donut_6_radial_outer_128.png");
+//    //auto donut = read_png_donut<double>("assets/donuts/crop/donut_6_radial_outer_256.png");
+
+//    //size_t steps = 200;
+//    //std::vector<double> sigma = {50,30,15,5};
+//    //size_t steps = 1000; // 20;
+//    std::vector<double> sigma = {2}; //{1.75}; //{2, 1.25}; //{5, 2}; // {15, 2}
+
+//    double init_x = donut.columns() / 2;
+//    double init_y = donut.rows() / 2;
+//    double init_r = donut.columns() / 3;
+
+
+////    std::cout << "initial ellipse position:\n xc = " << init_x << " yc = " << init_y << " a = " << init_r << " b = " << init_r
+////              << " phi = " << 0 << std::endl;
+
+//    //* // padding & gradient tests
+////    blaze::DynamicMatrix<double> A {{1, 2, 3, 4, 5}, {6, 7, 8, 9, 10}, {11, 12, 13, 14, 15}, {16, 17, 18, 19, 20}};
+////    //blaze::DynamicMatrix<double> A {{0, 0, 0, 0}, {0, 1, 0, 0}, {0, 0, 0, 0}};
+
+////    auto B = metric::DPM_detail::addPad(A);
+////    std::cout << A << "\n";
+////    std::cout << B << "\n";
+////    //auto G = metric::DPM_detail::gradient(A); //B);
+////    //std::cout << std::get<0>(G) << "\n" << std::get<1>(G) << "\n";
+////    metric::DPM_detail::updatePad(A);
+////    std::cout << A << "\n";
+////    metric::DPM_detail::updatePad(B);
+////    std::cout << B << "\n";
+////    blaze::submatrix(B, 1, 1, 4, 5) = {{51, 52, 53, 54, 55}, {56, 57, 58, 59, 60}, {61, 62, 63, 64, 65}, {66, 67, 68, 69, 70}};
+////    std::cout << B << "\n";
+////    metric::DPM_detail::updatePad(B);
+////    std::cout << B << "\n";
+////    auto C = metric::DPM_detail::removePad(B);
+////    std::cout << C << "\n";
+//    // */
+
+//    /* // GVF test
+
+//    blaze::DynamicMatrix<double> I = {
+//        {1, 0, 0, 0, 0 ,0, 0, 0, 0, 1},
+//        {0, 8, 2, 0, 0 ,0, 0, 0, 7, 0},
+//        {0, 0, 2, 0, 0 ,0, 0, 0, 0, 0},
+//        {0, 0, 2, 5, 0 ,0, 0, 0, 0, 0},
+//        {0, 0, 3, 6, 2 ,0, 0, 0, 0, 0},
+//        {1, 0, 2, 4, 1 ,0, 0, 1, 0, 0},
+//        {0, 0, 2, 0, 0 ,0, 1, 2, 1, 0},
+//        {0, 0, 1, 0, 0 ,1, 2, 3, 2, 1},
+//        {0, 5, 0, 0, 0 ,0, 1, 2, 6, 0},
+//        {1, 0, 2, 0, 0 ,0, 0, 1, 0, 1}
+//    };
+
+//    auto [h_1, v_1] = metric::DPM_detail::gvf(I, 0.1, 1, 10);
+
+//    std::cout << h_1 << "\n" << v_1 << "\n";
+//    return 0;
+
+////            // original code
+////            blaze::DynamicMatrix<double> f(I);
+
+////            size_t m = f.rows();
+////            size_t n = f.columns();
+
+////            // normalization
+////            auto fmin = blaze::min(f);
+////            auto fmax = blaze::max(f);
+
+////            if (fmax <= fmin) {
+////                std::cout << "Error: constant Input Matrix." << std::endl;
+////            }
+////            for (size_t i = 0; i < m; i++) {
+////                for (size_t j = 0; j < n; j++) {
+////                    f(i, j) = (f(i, j) - fmin) / (fmax - fmin);
+////                }
+////            }
+
+////    auto f2 = metric::DPM_detail::addPad(f);
+
+////    //std::cout << f2 << "\n";
+
+////    auto [fx, fy] = metric::DPM_detail::gradient(f2);
+
+////    //std::cout << fx << "\n" << fy << "\n";
+
+////            // originl code
+////            // square of magnitude
+////            blaze::DynamicMatrix<double> fxy_square((m + 2), (n + 2));
+////            for (size_t i = 0; i < m + 2; i++) {
+////                for (size_t j = 0; j < n + 2; ++j) {
+////                    fxy_square(i, j) = fx(i, j) * fx(i, j) + fy(i, j) * fy(i, j);
+////                }
+////            }
+
+////            blaze::DynamicMatrix<double> u1(fx);
+////            blaze::DynamicMatrix<double> v1(fy);
+////            //blaze::DynamicMatrix<double> Lu1((m + 2), (n + 2));
+////            //blaze::DynamicMatrix<double> Lv1((m + 2), (n + 2));
+////            blaze::DynamicMatrix<double> Lu1;//((m + 2), (n + 2), 0);
+////            blaze::DynamicMatrix<double> Lv1;//((m + 2), (n + 2), 0);
+
+////            //for (size_t it = 0; it < iter; it++) {
+////            for (size_t it = 0; it < 1; it++) {
+////                metric::DPM_detail::updatePad(u1);
+////                metric::DPM_detail::updatePad(v1);
+////                Lu1 = metric::DPM_detail::laplacian(u1);
+////                Lv1 = metric::DPM_detail::laplacian(v1);
+
+////                for (size_t i = 0; i < (m + 2); i++) {
+////                    for (size_t j = 0; j < (n + 2); ++j) {
+////                        u1(i, j) = u1(i, j) + alpha * (mu * Lu1(i, j) - fxy_square(i, j) * (u1(i, j) - fx(i, j)));
+////                        v1(i, j) = v1(i, j) + alpha * (mu * Lv1(i, j) - fxy_square(i, j) * (v1(i, j) - fy(i, j)));
+
+////                    }
+////                }
+////            }
+
+////    //std::cout << fxy_square << "\n";
+////    std::cout << u1 << "\n";
+////    std::cout << v1 << "\n";
+
+//    // */
+
+
+//    /* // mask test
 //    auto mask = weightingMask<double>(donut.rows(), donut.columns(), donut.columns()/3, 6);
-//    donut = mask % donut;
+//    vector2bmp(matrix2vv(mask), "mask_outer.bmp");
+//    //mask = weightingMask<double>(128, 128, 10, 2);
+//    //vector2bmp(matrix2vv(mask), "mask_inner.bmp");
 
-    /* // generate & save ring
-    auto ring = weightingMask<double>(128 - 7, 128 - 7, 40, 6);
-    blaze::DynamicMatrix<double> donut (128, 128, 0);
-    blaze::submatrix(donut, 7, 7, 128 - 7, 128 - 7) = ring;
-    blaze_dm_to_csv(donut, "ring.csv");
-    vector2bmp(matrix2vv(donut), "ring.bmp");
-    // */
+//    blaze::DynamicMatrix<double> applied = mask % donut;
+//    vector2bmp(matrix2vv(applied), "mask_applied.bmp");
 
-    auto donut = read_csv_blaze<double>("test01.csv", ",");
+//    //std::cout << mask << "\n";
+//    // */
 
 
-    //vector2bmp(matrix2vv(mask), "mask_outer.bmp");
+//    //* // fit_ellipse & forces test
 
-    //mask = weightingMask<double>(128, 128, 10, 2);
-    //vector2bmp(matrix2vv(mask), "mask_inner.bmp");
+//    init_x = donut.columns() / 2;
+//    init_y = donut.rows() / 2;
+//    init_r = donut.columns() / 3;
 
-    //vector2bmp(matrix2vv(donut), "mask_applied.bmp");
-
-
-    //auto donut = read_png_donut<double>("assets/donuts/crop/crop_2020-07-27_16_23_01_776_donut1.png");
-
-    //auto donut = read_png_donut<double>("assets/donuts/crop/donut_6_radial_outer_128.png");
-    //auto donut = read_png_donut<double>("assets/donuts/crop/donut_6_radial_outer_256.png");
-
-    //size_t steps = 200;
-    //std::vector<double> sigma = {50,30,15,5};
-    //size_t steps = 1000; // 20;
-    std::vector<double> sigma = {2}; //{1.75}; //{2, 1.25}; //{5, 2}; // {15, 2}
-
-    double init_x = donut.columns() / 2;
-    double init_y = donut.rows() / 2;
-    double init_r = donut.columns() / 3;
+//    //size_t filtersize = round(sigma[0] * 6); // 3 sigma
+//    //metric::imfilter<double, 1, metric::FilterType::GAUSSIAN, metric::PadDirection::BOTH, metric::PadType::CONST> f(filtersize, filtersize, sigma[0]);
+//    //auto I1 = f(donut);
+//    //auto I1 = metric::DPM_detail::gaussianBlur(donut, sigma[0]);
 
 
-//    std::cout << "initial ellipse position:\n xc = " << init_x << " yc = " << init_y << " a = " << init_r << " b = " << init_r
-//              << " phi = " << 0 << std::endl;
+////    auto gk = metric::DPM_detail::gaussianKernel(sigma[0]);
+////    blaze::DynamicMatrix<double> I1 = metric::DPM_detail::blackPaddedConv(donut, gk);
+////    vector2bmp(matrix2vv(I1), "blurred.bmp");
+////    mat2bmp::blaze2bmp_norm(I1, "blurred_norm.bmp");
+////    //I1 = blaze::submatrix(I1, (I1.rows() - donut.rows()) / 2, (I1.columns() - donut.columns()) / 2, donut.rows(), donut.columns());
+////    //vector2bmp(matrix2vv(I1), "blurred_cropped.bmp");
 
-    //* // padding & gradient tests, TODO delete
-//    blaze::DynamicMatrix<double> A {{1, 2, 3, 4, 5}, {6, 7, 8, 9, 10}, {11, 12, 13, 14, 15}, {16, 17, 18, 19, 20}};
-//    //blaze::DynamicMatrix<double> A {{0, 0, 0, 0}, {0, 1, 0, 0}, {0, 0, 0, 0}};
+////    std::cout << "blur input: min: " << blaze::min(donut) << ", max: " << blaze::max(donut) << "\n";
+////    std::cout << "GVF input: min: " << blaze::min(I1) << ", max: " << blaze::max(I1) << "\n";
 
-//    auto B = metric::DPM_detail::addPad(A);
-//    std::cout << A << "\n";
-//    std::cout << B << "\n";
-//    //auto G = metric::DPM_detail::gradient(A); //B);
-//    //std::cout << std::get<0>(G) << "\n" << std::get<1>(G) << "\n";
-//    metric::DPM_detail::updatePad(A);
-//    std::cout << A << "\n";
-//    metric::DPM_detail::updatePad(B);
-//    std::cout << B << "\n";
-//    blaze::submatrix(B, 1, 1, 4, 5) = {{51, 52, 53, 54, 55}, {56, 57, 58, 59, 60}, {61, 62, 63, 64, 65}, {66, 67, 68, 69, 70}};
-//    std::cout << B << "\n";
-//    metric::DPM_detail::updatePad(B);
-//    std::cout << B << "\n";
-//    auto C = metric::DPM_detail::removePad(B);
-//    std::cout << C << "\n";
-    // */
 
-    /* // GVF test, TODO remove
+//    auto I1 = donut;
 
-    blaze::DynamicMatrix<double> I = {
-        {1, 0, 0, 0, 0 ,0, 0, 0, 0, 1},
-        {0, 8, 2, 0, 0 ,0, 0, 0, 7, 0},
-        {0, 0, 2, 0, 0 ,0, 0, 0, 0, 0},
-        {0, 0, 2, 5, 0 ,0, 0, 0, 0, 0},
-        {0, 0, 3, 6, 2 ,0, 0, 0, 0, 0},
-        {1, 0, 2, 4, 1 ,0, 0, 1, 0, 0},
-        {0, 0, 2, 0, 0 ,0, 1, 2, 1, 0},
-        {0, 0, 1, 0, 0 ,1, 2, 3, 2, 1},
-        {0, 5, 0, 0, 0 ,0, 1, 2, 6, 0},
-        {1, 0, 2, 0, 0 ,0, 0, 1, 0, 1}
-    };
+//    auto [u1, v1] = metric::DPM_detail::gvf(I1, 0.1, 1, 10);
+//    vector2bmp(matrix2vv(u1), "u1.bmp");
+//    vector2bmp(matrix2vv(v1), "v1.bmp");
 
-    auto [h_1, v_1] = metric::DPM_detail::gvf(I, 0.1, 1, 10);
+//    u1 = read_csv_blaze<double>("u.csv", ","); // TODO remove
+//    v1 = read_csv_blaze<double>("v.csv", ",");
+//    vector2bmp(matrix2vv(u1), "u2.bmp");
+//    vector2bmp(matrix2vv(v1), "v2.bmp");
 
-    std::cout << h_1 << "\n" << v_1 << "\n";
-    return 0;
+//    std::cout << "input: " << blaze::min(I1) << ".." << blaze::max(I1) << "\n";
+//    std::cout << "u1: " << blaze::min(u1) << ".." << blaze::max(u1) << "\n";
+//    std::cout << "v1: " << blaze::min(v1) << ".." << blaze::max(v1) << "\n";
 
-//            // original code
-//            blaze::DynamicMatrix<double> f(I);
+//    //blaze_dm_to_csv(I1, "test01_read.csv");
+//    //blaze_dm_to_csv(u1, "u1.csv");
+//    //blaze_dm_to_csv(v1, "v1.csv");
 
-//            size_t m = f.rows();
-//            size_t n = f.columns();
+//    //std::cout << "\ninput: \n" << donut << "\n";
+//    //std::cout << "\nblurred: \n" << I1 << "\n";
+//    //std::cout << "\nforse field: \n" << u1 << "\n" << v1 << "\n";
 
-//            // normalization
-//            auto fmin = blaze::min(f);
-//            auto fmax = blaze::max(f);
+//    std::vector<double> ep = { init_x, init_y, init_r, init_r, 0 };  // initial parameter guess
+//    double thresh = 1e-6;
+//    double incr = 0.2;
+//    blaze::DynamicVector<double> increment = { incr, incr, incr, incr, M_PI / 180 * incr };  // increment in each iteration
+//    blaze::DynamicVector<double> threshold = { thresh, thresh, thresh, thresh, thresh };  // threshold for forces/torsinal moments
+//    double half_min_size = (donut.rows() < donut.columns() ? donut.rows() : donut.columns()) / 2.0;
+//    std::vector<double> bound = { 5, half_min_size, 5, half_min_size };  // the lower/upper bounds of a and b
 
-//            if (fmax <= fmin) {
-//                std::cout << "Error: constant Input Matrix." << std::endl;
+//    std::cout << "\nfit_ellipse test:\n";
+
+////    ep = metric::DPM_detail::fit_ellipse(ep, sigma[0] / 5 * increment, sigma[0] / 5 * threshold, bound, u1, v1, steps / sigma.size());
+////    for (size_t i = 0; i < ep.size(); ++i) {
+////        std::cout << ep[i] << "\t";
+////    }
+////    std::cout << "\n  by parts:\n";
+
+
+//    size_t m = u1.rows();
+//    size_t n = u1.columns();
+
+//    double xc = init_x;
+//    double yc = init_y;
+//    double a = init_r;
+//    double b = init_r;
+//    double phi = 0;
+
+//    increment = sigma[0] / 5 * increment;
+//    threshold = sigma[0] / 5 * threshold;
+
+//    size_t local_n_iter = 2000;
+
+//    for (size_t it = 0; it < local_n_iter; ++it) {
+
+//        // compute grid points from ellipse parameter
+//        std::vector<blaze::DynamicVector<double>> x_y_theta = metric::DPM_detail::ellipse2grid(m, n, xc, yc, a, b, phi);
+
+//        // torsion along the ellpise about center
+//        double torsion = metric::DPM_detail::torsion_moment(u1, v1, x_y_theta[0], x_y_theta[1], x_y_theta[2], xc, yc, phi);
+
+//        //std::cout << "theta: \n" << x_y_theta[0] << "\n" << x_y_theta[1] << "\n" << x_y_theta[2] << "\n";
+//        //std::cout << "gvf_x: min: " << blaze::min(u1) << ", max: " << blaze::max(u1) << "\n";
+//        //std::cout << "gvf_y: min: " << blaze::min(v1) << ", max: " << blaze::max(v1) << "\n";
+
+
+//        // update phi
+//        if (torsion > threshold[4]) {
+//            phi = phi + increment[4];
+//        }
+//        if (torsion < -threshold[4]) {
+//            phi = phi - increment[4];
+//        }
+
+//        std::cout << "torsion_moment: \n" << torsion << "\n";
+
+//        // F_around
+//        blaze::DynamicMatrix<double> iresult = metric::DPM_detail::contourForces(u1, v1, x_y_theta[0], x_y_theta[1]);
+//        blaze::DynamicVector<double, blaze::rowVector> F_round = blaze::sum<blaze::columnwise>(iresult);
+
+//        for (size_t i = 0; i < F_round.size(); ++i) {
+//            F_round[i] = F_round[i] / double(x_y_theta[2].size());
+//        }
+
+//        std::vector<double> Fround(F_round.size());
+//        for (size_t i = 0; i < F_round.size(); ++i) {
+//            Fround[i] = F_round[i];
+//        }
+
+//        std::cout << "f_round: \n" << F_round << "\n";
+
+
+//        std::vector<double> x_index1;
+//        std::vector<double> y_index1;
+//        std::vector<double> x_index2;
+//        std::vector<double> y_index2;
+//        std::vector<double> x_index3;
+//        std::vector<double> y_index3;
+//        std::vector<double> x_index4;
+//        std::vector<double> y_index4;
+
+//        for (size_t i = 0; i < x_y_theta[0].size(); ++i) {
+//            if (x_y_theta[2][i] > M_PI * 3 / 4 && x_y_theta[2][i] < M_PI * 5 / 4) {
+//                x_index1.push_back(x_y_theta[0][i]);
+//                y_index1.push_back(x_y_theta[1][i]);
 //            }
-//            for (size_t i = 0; i < m; i++) {
-//                for (size_t j = 0; j < n; j++) {
-//                    f(i, j) = (f(i, j) - fmin) / (fmax - fmin);
-//                }
+//            if (x_y_theta[2][i] < M_PI / 4 || x_y_theta[2][i] > M_PI * 7 / 4) {
+//                x_index2.push_back(x_y_theta[0][i]);
+//                y_index2.push_back(x_y_theta[1][i]);
 //            }
-
-//    auto f2 = metric::DPM_detail::addPad(f);
-
-//    //std::cout << f2 << "\n";
-
-//    auto [fx, fy] = metric::DPM_detail::gradient(f2);
-
-//    //std::cout << fx << "\n" << fy << "\n";
-
-//            // originl code
-//            // square of magnitude
-//            blaze::DynamicMatrix<double> fxy_square((m + 2), (n + 2));
-//            for (size_t i = 0; i < m + 2; i++) {
-//                for (size_t j = 0; j < n + 2; ++j) {
-//                    fxy_square(i, j) = fx(i, j) * fx(i, j) + fy(i, j) * fy(i, j);
-//                }
+//            if (x_y_theta[2][i] > M_PI / 4 && x_y_theta[2][i] < M_PI * 3 / 4) {
+//                x_index3.push_back(x_y_theta[0][i]);
+//                y_index3.push_back(x_y_theta[1][i]);
 //            }
-
-//            blaze::DynamicMatrix<double> u1(fx);
-//            blaze::DynamicMatrix<double> v1(fy);
-//            //blaze::DynamicMatrix<double> Lu1((m + 2), (n + 2));
-//            //blaze::DynamicMatrix<double> Lv1((m + 2), (n + 2));
-//            blaze::DynamicMatrix<double> Lu1;//((m + 2), (n + 2), 0);
-//            blaze::DynamicMatrix<double> Lv1;//((m + 2), (n + 2), 0);
-
-//            //for (size_t it = 0; it < iter; it++) {
-//            for (size_t it = 0; it < 1; it++) {
-//                metric::DPM_detail::updatePad(u1);
-//                metric::DPM_detail::updatePad(v1);
-//                Lu1 = metric::DPM_detail::laplacian(u1);
-//                Lv1 = metric::DPM_detail::laplacian(v1);
-
-//                for (size_t i = 0; i < (m + 2); i++) {
-//                    for (size_t j = 0; j < (n + 2); ++j) {
-//                        u1(i, j) = u1(i, j) + alpha * (mu * Lu1(i, j) - fxy_square(i, j) * (u1(i, j) - fx(i, j)));
-//                        v1(i, j) = v1(i, j) + alpha * (mu * Lv1(i, j) - fxy_square(i, j) * (v1(i, j) - fy(i, j)));
-
-//                    }
-//                }
+//            if (x_y_theta[2][i] > M_PI * 5 / 4 && x_y_theta[2][i] < M_PI * 7 / 4) {
+//                x_index4.push_back(x_y_theta[0][i]);
+//                y_index4.push_back(x_y_theta[1][i]);
 //            }
+//        }
 
-//    //std::cout << fxy_square << "\n";
-//    std::cout << u1 << "\n";
-//    std::cout << v1 << "\n";
+//        double F_left = metric::DPM_detail::force(u1, v1, x_index1, y_index1, std::cos(phi), std::sin(phi));
+//        double F_right = metric::DPM_detail::force(u1, v1, x_index2, y_index2, -std::cos(phi), -std::sin(phi));
+//        double F_up = metric::DPM_detail::force(u1, v1, x_index3, y_index3, std::sin(phi), -std::cos(phi));
+//        double F_down = metric::DPM_detail::force(u1, v1, x_index4, y_index4, -std::sin(phi), std::cos(phi));
 
-    // */
+//        // std::cout << "it=" << it << " --> " << F_left << " " << F_right << " " << F_up << " " << F_down
+//        //           << std::endl;
 
+//        // update xc and yc
+//        double F_left_right = F_round[0] * 1 + F_round[1] * 0;
+//        if (F_left_right > threshold[0]) {
+//            xc = xc + increment[0];
+//            ;
+//        } else if (F_left_right < -threshold[0]) {
+//            xc = xc - increment[0];
+//        }
 
-    /* // mask test, TODO delete
-    auto mask = weightingMask<double>(donut.rows(), donut.columns(), donut.columns()/3, 6);
-    vector2bmp(matrix2vv(mask), "mask_outer.bmp");
-    //mask = weightingMask<double>(128, 128, 10, 2);
-    //vector2bmp(matrix2vv(mask), "mask_inner.bmp");
+//        double F_down_up = F_round[0] * 0 + F_round[1] * 1;
+//        if (F_down_up > threshold[1]) {
+//            yc = yc + increment[1];
+//        } else if (F_down_up < -threshold[1]) {
+//            yc = yc - increment[1];
+//        }
 
-    blaze::DynamicMatrix<double> applied = mask % donut;
-    vector2bmp(matrix2vv(applied), "mask_applied.bmp");
+//        // update xc and yc again according to diagonal force
+//        double F_diag1 = F_round[0] * 0.7071 + F_round[1] * 0.7071;
+//        if (F_diag1 > threshold[0] + threshold[1]) {
+//            xc = xc + increment[0];
+//            yc = yc + increment[1];
+//        } else if (F_diag1 < -threshold[0] - threshold[1]) {
+//            xc = xc - increment[0];
+//            yc = yc - increment[1];
+//        }
 
-    //std::cout << mask << "\n";
-    // */
+//        double F_diag2 = F_round[0] * (-0.7071) + F_round[1] * 0.7071;
+//        if (F_diag2 > threshold[0] + threshold[1]) {
+//            xc = xc - increment[0];
+//            yc = yc + increment[1];
+//        } else if (F_diag2 < -threshold[0] - threshold[1]) {
+//            xc = xc + increment[0];
+//            yc = yc - increment[1];
+//        }
 
+//        // update a and b
 
-    //* // fit_ellipse & forces test, TODO delete
+//        if (F_left + F_right > threshold[2])
+//            a = a - increment[2];
+//        else if (F_left + F_right < -threshold[2])
+//            a = a + increment[2];
 
-    init_x = donut.columns() / 2;
-    init_y = donut.rows() / 2;
-    init_r = donut.columns() / 3;
+//        if (F_up + F_down > threshold[3])
+//            b = b - increment[3];
+//        else if (F_up + F_down < -threshold[3])
+//            b = b + increment[3];
 
-    //size_t filtersize = round(sigma[0] * 6); // 3 sigma
-    //metric::imfilter<double, 1, metric::FilterType::GAUSSIAN, metric::PadDirection::BOTH, metric::PadType::CONST> f(filtersize, filtersize, sigma[0]);
-    //auto I1 = f(donut);
-    //auto I1 = metric::DPM_detail::gaussianBlur(donut, sigma[0]);
+//        if (b > a) {
+//            std::swap(a, b);
+//            phi = std::fmod(phi + M_PI / 2, M_PI);
+//        }
 
+//        // restrict a and b using lower and upper bounds
+//        if (a > bound[1])
+//            a = bound[1];
 
-//    auto gk = metric::DPM_detail::gaussianKernel(sigma[0]);
-//    blaze::DynamicMatrix<double> I1 = metric::DPM_detail::blackPaddedConv(donut, gk);
-//    vector2bmp(matrix2vv(I1), "blurred.bmp");
-//    mat2bmp::blaze2bmp_norm(I1, "blurred_norm.bmp");
-//    //I1 = blaze::submatrix(I1, (I1.rows() - donut.rows()) / 2, (I1.columns() - donut.columns()) / 2, donut.rows(), donut.columns());
-//    //vector2bmp(matrix2vv(I1), "blurred_cropped.bmp");
+//        if (a < bound[0])
+//            a = bound[0];
 
-//    std::cout << "blur input: min: " << blaze::min(donut) << ", max: " << blaze::max(donut) << "\n";
-//    std::cout << "GVF input: min: " << blaze::min(I1) << ", max: " << blaze::max(I1) << "\n";
+//        if (b > bound[3])
+//            b = bound[3];
 
+//        if (b < bound[2])
+//            b = bound[2];
 
-    auto I1 = donut;
+//        std::cout << "iter result: " << xc << " " << yc << " " << a << " " << b << " " << phi << " | " << it << "\n\n";  // TODO remove
 
-    auto [u1, v1] = metric::DPM_detail::gvf(I1, 0.1, 1, 10);
-    vector2bmp(matrix2vv(u1), "u1.bmp");
-    vector2bmp(matrix2vv(v1), "v1.bmp");
-
-    u1 = read_csv_blaze<double>("u.csv", ","); // TODO remove
-    v1 = read_csv_blaze<double>("v.csv", ",");
-    vector2bmp(matrix2vv(u1), "u2.bmp");
-    vector2bmp(matrix2vv(v1), "v2.bmp");
-
-    std::cout << "input: " << blaze::min(I1) << ".." << blaze::max(I1) << "\n";
-    std::cout << "u1: " << blaze::min(u1) << ".." << blaze::max(u1) << "\n";
-    std::cout << "v1: " << blaze::min(v1) << ".." << blaze::max(v1) << "\n";
-
-    //blaze_dm_to_csv(I1, "test01_read.csv");
-    //blaze_dm_to_csv(u1, "u1.csv");
-    //blaze_dm_to_csv(v1, "v1.csv");
-
-    //std::cout << "\ninput: \n" << donut << "\n";
-    //std::cout << "\nblurred: \n" << I1 << "\n";
-    //std::cout << "\nforse field: \n" << u1 << "\n" << v1 << "\n";
-
-    std::vector<double> ep = { init_x, init_y, init_r, init_r, 0 };  // initial parameter guess
-    double thresh = 1e-6;
-    double incr = 0.2;
-    blaze::DynamicVector<double> increment = { incr, incr, incr, incr, M_PI / 180 * incr };  // increment in each iteration
-    blaze::DynamicVector<double> threshold = { thresh, thresh, thresh, thresh, thresh };  // threshold for forces/torsinal moments
-    double half_min_size = (donut.rows() < donut.columns() ? donut.rows() : donut.columns()) / 2.0;
-    std::vector<double> bound = { 5, half_min_size, 5, half_min_size };  // the lower/upper bounds of a and b
-
-    std::cout << "\nfit_ellipse test:\n";
-
-//    ep = metric::DPM_detail::fit_ellipse(ep, sigma[0] / 5 * increment, sigma[0] / 5 * threshold, bound, u1, v1, steps / sigma.size());
-//    for (size_t i = 0; i < ep.size(); ++i) {
-//        std::cout << ep[i] << "\t";
 //    }
-//    std::cout << "\n  by parts:\n";
 
+//    // */
 
-    size_t m = u1.rows();
-    size_t n = u1.columns();
-
-    double xc = init_x;
-    double yc = init_y;
-    double a = init_r;
-    double b = init_r;
-    double phi = 0;
-
-    increment = sigma[0] / 5 * increment;
-    threshold = sigma[0] / 5 * threshold;
-
-    size_t local_n_iter = 2000;
-
-    for (size_t it = 0; it < local_n_iter; ++it) {
-
-        // compute grid points from ellipse parameter
-        std::vector<blaze::DynamicVector<double>> x_y_theta = metric::DPM_detail::ellipse2grid(m, n, xc, yc, a, b, phi);
-
-        // torsion along the ellpise about center
-        double torsion = metric::DPM_detail::torsion_moment(u1, v1, x_y_theta[0], x_y_theta[1], x_y_theta[2], xc, yc, phi);
-
-        //std::cout << "theta: \n" << x_y_theta[0] << "\n" << x_y_theta[1] << "\n" << x_y_theta[2] << "\n";
-        //std::cout << "gvf_x: min: " << blaze::min(u1) << ", max: " << blaze::max(u1) << "\n";
-        //std::cout << "gvf_y: min: " << blaze::min(v1) << ", max: " << blaze::max(v1) << "\n";
-
-
-        // update phi
-        if (torsion > threshold[4]) {
-            phi = phi + increment[4];
-        }
-        if (torsion < -threshold[4]) {
-            phi = phi - increment[4];
-        }
-
-        std::cout << "torsion_moment: \n" << torsion << "\n";
-
-        // F_around
-        blaze::DynamicMatrix<double> iresult = metric::DPM_detail::contourForces(u1, v1, x_y_theta[0], x_y_theta[1]);
-        blaze::DynamicVector<double, blaze::rowVector> F_round = blaze::sum<blaze::columnwise>(iresult);
-
-        for (size_t i = 0; i < F_round.size(); ++i) {
-            F_round[i] = F_round[i] / double(x_y_theta[2].size());
-        }
-
-        std::vector<double> Fround(F_round.size());
-        for (size_t i = 0; i < F_round.size(); ++i) {
-            Fround[i] = F_round[i];
-        }
-
-        std::cout << "f_round: \n" << F_round << "\n";
-
-
-        std::vector<double> x_index1;
-        std::vector<double> y_index1;
-        std::vector<double> x_index2;
-        std::vector<double> y_index2;
-        std::vector<double> x_index3;
-        std::vector<double> y_index3;
-        std::vector<double> x_index4;
-        std::vector<double> y_index4;
-
-        for (size_t i = 0; i < x_y_theta[0].size(); ++i) {
-            if (x_y_theta[2][i] > M_PI * 3 / 4 && x_y_theta[2][i] < M_PI * 5 / 4) {
-                x_index1.push_back(x_y_theta[0][i]);
-                y_index1.push_back(x_y_theta[1][i]);
-            }
-            if (x_y_theta[2][i] < M_PI / 4 || x_y_theta[2][i] > M_PI * 7 / 4) {
-                x_index2.push_back(x_y_theta[0][i]);
-                y_index2.push_back(x_y_theta[1][i]);
-            }
-            if (x_y_theta[2][i] > M_PI / 4 && x_y_theta[2][i] < M_PI * 3 / 4) {
-                x_index3.push_back(x_y_theta[0][i]);
-                y_index3.push_back(x_y_theta[1][i]);
-            }
-            if (x_y_theta[2][i] > M_PI * 5 / 4 && x_y_theta[2][i] < M_PI * 7 / 4) {
-                x_index4.push_back(x_y_theta[0][i]);
-                y_index4.push_back(x_y_theta[1][i]);
-            }
-        }
-
-        double F_left = metric::DPM_detail::force(u1, v1, x_index1, y_index1, std::cos(phi), std::sin(phi));
-        double F_right = metric::DPM_detail::force(u1, v1, x_index2, y_index2, -std::cos(phi), -std::sin(phi));
-        double F_up = metric::DPM_detail::force(u1, v1, x_index3, y_index3, std::sin(phi), -std::cos(phi));
-        double F_down = metric::DPM_detail::force(u1, v1, x_index4, y_index4, -std::sin(phi), std::cos(phi));
-
-        // std::cout << "it=" << it << " --> " << F_left << " " << F_right << " " << F_up << " " << F_down
-        //           << std::endl;
-
-        // update xc and yc
-        double F_left_right = F_round[0] * 1 + F_round[1] * 0;
-        if (F_left_right > threshold[0]) {
-            xc = xc + increment[0];
-            ;
-        } else if (F_left_right < -threshold[0]) {
-            xc = xc - increment[0];
-        }
-
-        double F_down_up = F_round[0] * 0 + F_round[1] * 1;
-        if (F_down_up > threshold[1]) {
-            yc = yc + increment[1];
-        } else if (F_down_up < -threshold[1]) {
-            yc = yc - increment[1];
-        }
-
-        // update xc and yc again according to diagonal force
-        double F_diag1 = F_round[0] * 0.7071 + F_round[1] * 0.7071;
-        if (F_diag1 > threshold[0] + threshold[1]) {
-            xc = xc + increment[0];
-            yc = yc + increment[1];
-        } else if (F_diag1 < -threshold[0] - threshold[1]) {
-            xc = xc - increment[0];
-            yc = yc - increment[1];
-        }
-
-        double F_diag2 = F_round[0] * (-0.7071) + F_round[1] * 0.7071;
-        if (F_diag2 > threshold[0] + threshold[1]) {
-            xc = xc - increment[0];
-            yc = yc + increment[1];
-        } else if (F_diag2 < -threshold[0] - threshold[1]) {
-            xc = xc + increment[0];
-            yc = yc - increment[1];
-        }
-
-        // update a and b
-
-        if (F_left + F_right > threshold[2])
-            a = a - increment[2];
-        else if (F_left + F_right < -threshold[2])
-            a = a + increment[2];
-
-        if (F_up + F_down > threshold[3])
-            b = b - increment[3];
-        else if (F_up + F_down < -threshold[3])
-            b = b + increment[3];
-
-        if (b > a) {
-            std::swap(a, b);
-            phi = std::fmod(phi + M_PI / 2, M_PI);
-        }
-
-        // restrict a and b using lower and upper bounds
-        if (a > bound[1])
-            a = bound[1];
-
-        if (a < bound[0])
-            a = bound[0];
-
-        if (b > bound[3])
-            b = bound[3];
-
-        if (b < bound[2])
-            b = bound[2];
-
-        std::cout << "iter result: " << xc << " " << yc << " " << a << " " << b << " " << phi << " | " << it << "\n\n";  // TODO remove
-
-    }
-
-    // */
-
-    return 0;
-}
+//    return 0;
+//}
