+++ examples/transform_examples/donuts.hpp
@@ -1,46 +1,17 @@
 #ifndef DONUTS_HPP
 #define DONUTS_HPP
 
-#include "../../modules/transform/distance_potential_minimization.hpp"
+#include "../../3rdparty/blaze/Blaze.h"
 
 #include <cmath>
-//#include <tuple>
-#include <boost/gil/extension/io/jpeg.hpp>
-#include <boost/gil/extension/io/png.hpp>
-//#include <boost/gil/extension/io/png_io.hpp>
-//#include <boost/gil/extension/io/png_dynamic_io.hpp>
-#include "../../3rdparty/blaze/Blaze.h" // for only read_png_donut
-//#include <opencv2/opencv.hpp>
 #include <iostream>
-#include <fstream>
+//#include <tuple>
+//#include <opencv2/opencv.hpp>
 
 
-
-// compile with -ljpeg -lopencv_core -lopencv_imgproc
 
 
 /*
-
-blaze::DynamicMatrix<double> cv2blaze(cv::Mat in) {
-    blaze::DynamicMatrix<double> out (in.rows, in.cols, 0);
-    for (int y = 0; y<in.rows; y++) {
-        for (int x = 0; x<in.cols; x++) {
-            out(y, x) = in.at<double>(y, x);
-        }
-    }
-    return out;
-}
-
-cv::Mat blaze2cv(blaze::DynamicMatrix<double> in) {
-    cv::Mat out (in.rows(), in.columns(), cv::DataType<double>::type);
-    for (std::size_t y = 0; y<in.rows(); y++) {
-        for (std::size_t x = 0; x<in.columns(); x++) {
-            out.at<double>(y, x) = in(y, x);
-        }
-    }
-    return out;
-}
-
 
 blaze::DynamicMatrix<double> create_filter(double sigma) {
     std::size_t size = ceil(sigma)*3 + 1;
@@ -88,98 +59,6 @@ blaze::DynamicMatrix<double> gaussianBlur(blaze::DynamicMatrix<double> in, doubl
 }
 
 // */
-
-
-template <class ContainerType>
-void vv_to_csv(ContainerType data, std::string filename, std::string sep=",")  // container
-{
-    std::ofstream outputFile;
-    outputFile.open(filename);
-        for (auto i = 0; i < data.size(); ++i) {
-            for (auto j = 0; j < data[i].size(); j++) {
-                outputFile << std::to_string(data[i][j]);
-                if (j < data[i].size() - 1)
-                    outputFile << sep;
-            }
-            outputFile << std::endl;
-        }
-        outputFile.close();
-} // TODO add return flag
-
-
-
-template <typename T>
-void vector2bmp(std::vector<std::vector<T>> m, std::string filename, double amplify = 1)
-{ // TODO combine with blaze2bmp
-
-    int h = m.size();
-    assert(h>0);
-    int w = m[0].size();
-
-    int x, y, r, g, b;
-
-    FILE *f;
-    unsigned char *img = NULL;
-    int filesize = 54 + 3*w*h;
-
-    img = (unsigned char *)malloc(3*w*h);
-    std::memset(img,0,3*w*h);
-
-    for(int i=0; i<w; i++)
-    {
-        for(int j=0; j<h; j++)
-        {
-            x=i; y=j;
-            r = 0;
-            g = 0;
-            b = 0;
-            int p = m[j][i]*255*amplify;
-            if (p > 0) // green for positive, red for negative
-            {
-                g = p;
-                b = p;
-            }
-            else
-                r = -p;
-            if (r > 255) r=255;
-            if (g > 255) g=255;
-            if (b > 255) b=255;
-            img[(x+y*w)*3+2] = (unsigned char)(r);
-            img[(x+y*w)*3+1] = (unsigned char)(g);
-            img[(x+y*w)*3+0] = (unsigned char)(b);
-        }
-    }
-
-    unsigned char bmpfileheader[14] = {'B','M', 0,0,0,0, 0,0, 0,0, 54,0,0,0};
-    unsigned char bmpinfoheader[40] = {40,0,0,0, 0,0,0,0, 0,0,0,0, 1,0, 24,0};
-    unsigned char bmppad[3] = {0,0,0};
-
-    bmpfileheader[ 2] = (unsigned char)(filesize    );
-    bmpfileheader[ 3] = (unsigned char)(filesize>> 8);
-    bmpfileheader[ 4] = (unsigned char)(filesize>>16);
-    bmpfileheader[ 5] = (unsigned char)(filesize>>24);
-
-    bmpinfoheader[ 4] = (unsigned char)(       w    );
-    bmpinfoheader[ 5] = (unsigned char)(       w>> 8);
-    bmpinfoheader[ 6] = (unsigned char)(       w>>16);
-    bmpinfoheader[ 7] = (unsigned char)(       w>>24);
-    bmpinfoheader[ 8] = (unsigned char)(       h    );
-    bmpinfoheader[ 9] = (unsigned char)(       h>> 8);
-    bmpinfoheader[10] = (unsigned char)(       h>>16);
-    bmpinfoheader[11] = (unsigned char)(       h>>24);
-
-    f = fopen(filename.c_str(),"wb");
-    fwrite(bmpfileheader,1,14,f);
-    fwrite(bmpinfoheader,1,40,f);
-    for(int i=0; i<h; i++)
-    {
-        fwrite(img+(w*(h-i-1)*3),3,w,f);
-        fwrite(bmppad,1,(4-(w*3)%4)%4,f);
-    }
-
-    free(img);
-    fclose(f);
-}
 
 
 
@@ -238,7 +117,7 @@ std::vector<std::vector<T>> otsu_binarize(std::vector<std::vector<T>> in, float 
             }
         }
     }
-    //assert(pixel_counter==height*width); // FIXME detecter pixel loss, TODO fix!!
+    //assert(pixel_counter==height*width); // FIXME detector pixel loss, TODO fix!!
 
     // finding Otsu threshold according to https://en.wikipedia.org/wiki/Otsu%27s_method
     double omega0 = 0; //(double)hist[0] / (double)pixel_counter;
@@ -434,8 +313,8 @@ blaze::DynamicMatrix<T> radial_diff(const blaze::DynamicMatrix<T> & in, double s
 
 
 
-//*
-template <typename T> // TODO debug!!
+/*
+template <typename T> // old crop version, TODO debug if needed
 std::tuple<std::vector<std::vector<T>>, std::vector<std::vector<T>>>
 split_donut(std::vector<std::vector<T>> in, float crop_share = 3.0/5.0) {
 
@@ -465,37 +344,6 @@ split_donut(std::vector<std::vector<T>> in, float crop_share = 3.0/5.0) {
 // */
 
 
-template <typename T>
-blaze::DynamicMatrix<T> read_png_donut(std::string filename) {
-    boost::gil::rgb8_image_t img;
-    boost::gil::read_image(filename, img, boost::gil::png_tag());
-    auto gray = boost::gil::color_converted_view<boost::gil::gray8_pixel_t>(const_view(img));
-
-    blaze::DynamicMatrix<T> p (gray.height(), gray.width());
-    for (int y=0; y<gray.height(); ++y) {
-        for (int x=0; x<gray.width(); ++x) {
-            p(y, x) = gray(x, y)/255.0;
-        }
-    }
-    return p;
-}
-
-
-
-template <typename T>
-std::vector<std::vector<T>> matrix2vv(const blaze::DynamicMatrix<T> & mat) {
-    std::vector<std::vector<double>> v;
-    for (size_t i = 0; i < mat.rows(); ++i) {
-        std::vector<double> line;
-        for (size_t j = 0; j < mat.columns(); ++j) {
-            line.push_back(mat(i, j));
-        }
-        v.push_back(line);
-    }
-    return v;
-}
-
-
 
 template <typename T>
 blaze::DynamicMatrix<T> weightingMask(size_t h, size_t w, T radius, T sigma) { // radius is where we expect the donut outline
@@ -514,8 +362,9 @@ blaze::DynamicMatrix<T> weightingMask(size_t h, size_t w, T radius, T sigma) { /
 }
 
 
+/*
 
-void eat_donut(int number) {
+void eat_donut(int number) {  // old test batch procewssing function
 
     boost::gil::rgb8_image_t img;
     //boost::gil::read_image("assets/" + std::to_string(number) + ".jpg", img, boost::gil::jpeg_tag());
@@ -600,7 +449,7 @@ void eat_donut(int number) {
 
 }
 
-
+// */
 
 
 //int main() {
